Optimize memory performance, defer LLM startup, enhance sensor managers, and upgrade Gradle/Kotlin/AGP dependencies

+++ app/src/main/java/com/neuromesh/crisis/domain/consensus/VotingStrategy.kt
@@ -46,7 +46,7 @@ class VotingStrategy @Inject constructor() {
             ConsensusResult.Disputed(
                 topCrisisType = dominantCrisis.key,
                 agreementRatio = crisisRatio.toFloat(),
-                avgConfidence = avgConfidence
+                avgConfidence = avgConfidence.toDouble()
             )
         }
     }
